+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Data Guard service support.
+- Prompt Guard service support.
+
 ## 5.1.0 - 2024-10-16
 
 ### Added